fixed race condition error with chat after auto join room
