+++ bsg-frontend/apps/extension/hooks/useRoomEvents.ts
@@ -1,13 +1,10 @@
 import { useRoomStore } from "@/stores/useRoomStore";
-import { useChatSocket } from "./useChatSocket";
 import { SERVER_URL } from '../lib/config'
 import { useEffect } from "react";
 import { useUserStore } from "@/stores/useUserStore";
-import { User } from "@bsg/models/User";
 
 export function useRoomEvents() {
 
-    const { joinChatRoom } = useChatSocket();
     const isLoggedIn = useUserStore(s => s.isLoggedIn);
     const isInRoom = useRoomStore(s => s.isInRoom);
     const userId = useUserStore(s => s.userId);
@@ -17,8 +14,6 @@ export function useRoomEvents() {
     const setNextProblem = useRoomStore(s => s.setNextProblem);
     const setRoundEndTime = useRoomStore(s => s.setRoundEndTime);
     const setIsRoundStarted = useRoomStore(s => s.setIsRoundStarted);
-    const loginUser = useUserStore(s => s.loginUser);
-    const initRoom = useRoomStore(s => s.initRoom);
     const resetRoom = useRoomStore(s => s.resetRoom);
 
     // Handle Game Events
@@ -88,7 +83,6 @@ export function useRoomEvents() {
         } else if (lastGameEvent.type === 'round-end') {
             setRoundEndTime(null);
             setIsRoundStarted(false);
-            resetRoom();
 
             // Clear nextProblem and TTL state on round end
             setNextProblem(null);
@@ -172,7 +166,7 @@ export function useRoomEvents() {
                 setTimeout(() => {
                     setIsRoundStarted(false);
                     setRoundEndTime(null);
-                    resetRoom();
+                    
                     if (typeof chrome !== 'undefined' && chrome.storage && chrome.storage.local) {
                         chrome.storage.local.remove('roundEndTime');
                     }
@@ -184,83 +178,5 @@ export function useRoomEvents() {
         }
     }
 
-    useEffect(() => {
-        if (typeof chrome !== 'undefined' && chrome.runtime?.sendMessage) {
-            chrome.runtime.sendMessage({ type: 'CHECK_AUTH' }, (response) => {
-                if (response?.success) {
-                    const user: User = response.user;
-                    loginUser(
-                        user.id,
-                        user.name,
-                        user.email,
-                        user.photo
-                    );
-
-                    checkActiveRoom();
-                }
-            })
-        }
-    }, [])
-
-    const checkActiveRoom = async () => {
-        try {
-            const res = await fetch(`${SERVER_URL}/rooms/active`, { credentials: 'include' });
-            if (res.ok) {
-                const data = await res.json();
-                if (data.id || data.roomID) { // handle potentially different response structure
-                    const roomId = data.id || data.roomID;
-                    // Fetch room details to get round status
-                    const roomRes = await fetch(`${SERVER_URL}/rooms/${roomId}`, { credentials: 'include' });
-                    if (roomRes.ok) {
-                        const roomData = await roomRes.json();
-                        const room = roomData.data;
-                        console.log("CheckActiveRoom: Fetched room details", room);
-                        initRoom(
-                            room.id,
-                            room.shortCode,
-                            room.adminId,
-                            userId === room.adminId
-                        );
-
-                        if (typeof chrome !== 'undefined' && chrome.storage && chrome.storage.local) {
-                            console.log("CheckActiveRoom: Saving activeRoomId to storage", room.id);
-                            chrome.storage.local.set({ activeRoomId: room.id }, () => {
-                                console.log("CheckActiveRoom: Saved activeRoomId");
-                            });
-                        } else {
-                            console.warn("CheckActiveRoom: chrome.storage.local not available");
-                        }
-
-                        joinChatRoom(room.id);
-
-                        // Check for active round
-                        console.log("CheckActiveRoom: Rounds:", room.rounds);
-                        if (room.rounds && room.rounds.length > 0) {
-                            const lastRound = room.rounds[room.rounds.length - 1];
-                            const status = lastRound.Status || lastRound.status;
-                            console.log("CheckActiveRoom: Last round status:", status);
-                            // ROUND_STARTED = "started" (need to verify constant value, assuming string)
-                            if (status === "started") {
-                                setIsRoundStarted(true);
-                                const startTimeStr = lastRound.LastUpdatedTime || lastRound.lastUpdatedTime;
-                                const startTime = new Date(startTimeStr).getTime();
-                                const duration = lastRound.duration || lastRound.Duration;
-                                const endTime = startTime + (duration * 60 * 1000);
-                                if (endTime > Date.now()) {
-                                    setRoundEndTime(endTime);
-                                    if (typeof chrome !== 'undefined' && chrome.storage && chrome.storage.local) {
-                                        chrome.storage.local.set({ roundEndTime: endTime });
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        } catch (e) {
-            console.error("Failed to check active room", e);
-        }
-    }
-
-    return { handleStartRound, handleEndRound, checkActiveRoom };
+    return { handleStartRound, handleEndRound };
 }
