+++ bsg-frontend/apps/extension/hooks/useRoomInit.ts
@@ -10,6 +10,8 @@ export const useRoomInit = () => {
 
     const { joinChatRoom } = useChatSocket();
     const initRoom = useRoomStore(s => s.initRoom);
+    const setIsRoundStarted = useRoomStore(s => s.setIsRoundStarted);
+    const setRoundEndTime = useRoomStore(s => s.setRoundEndTime);
     const userId = useUserStore(s => s.userId);
 
     const joinRoom = async (roomCode: string) => {
@@ -111,5 +113,67 @@ export const useRoomInit = () => {
         }
     }
 
-    return { joinRoom, createRoom };
+    // TODO: Return a boolean to determine if room-choice is loaded
+    const checkActiveRoom = async () => {
+        try {
+            const res = await fetch(`${SERVER_URL}/rooms/active`, { credentials: 'include' });
+            if (res.ok) {
+                const data = await res.json();
+                if (data.id || data.roomID) { // handle potentially different response structure
+                    const roomId = data.id || data.roomID;
+                    // Fetch room details to get round status
+                    const roomRes = await fetch(`${SERVER_URL}/rooms/${roomId}`, { credentials: 'include' });
+                    if (roomRes.ok) {
+                        const roomData = await roomRes.json();
+                        const room = roomData.data;
+                        console.log("CheckActiveRoom: Fetched room details", room);
+                        initRoom(
+                            room.id,
+                            room.shortCode,
+                            room.adminId,
+                            userId === room.adminId
+                        );
+
+                        if (typeof chrome !== 'undefined' && chrome.storage && chrome.storage.local) {
+                            console.log("CheckActiveRoom: Saving activeRoomId to storage", room.id);
+                            chrome.storage.local.set({ activeRoomId: room.id }, () => {
+                                console.log("CheckActiveRoom: Saved activeRoomId");
+                            });
+                        } else {
+                            console.warn("CheckActiveRoom: chrome.storage.local not available");
+                        }
+
+                        joinChatRoom(room.id);
+                        router.push('/chat-page');
+
+                        // Check for active round
+                        console.log("CheckActiveRoom: Rounds:", room.rounds);
+                        if (room.rounds && room.rounds.length > 0) {
+                            const lastRound = room.rounds[room.rounds.length - 1];
+                            const status = lastRound.Status || lastRound.status;
+                            console.log("CheckActiveRoom: Last round status:", status);
+                            // ROUND_STARTED = "started" (need to verify constant value, assuming string)
+                            if (status === "started") {
+                                setIsRoundStarted(true);
+                                const startTimeStr = lastRound.LastUpdatedTime || lastRound.lastUpdatedTime;
+                                const startTime = new Date(startTimeStr).getTime();
+                                const duration = lastRound.duration || lastRound.Duration;
+                                const endTime = startTime + (duration * 60 * 1000);
+                                if (endTime > Date.now()) {
+                                    setRoundEndTime(endTime);
+                                    if (typeof chrome !== 'undefined' && chrome.storage && chrome.storage.local) {
+                                        chrome.storage.local.set({ roundEndTime: endTime });
+                                    }
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        } catch (e) {
+            console.error("Failed to check active room", e);
+        }
+    }
+
+    return { joinRoom, createRoom, checkActiveRoom };
 }
