+++ bsg-frontend/apps/extension/stores/usePanelStore.ts
@@ -1,4 +1,3 @@
-import { isatty } from 'tty';
 import { create } from 'zustand';
 
 interface panelStoreState {
